+++ mycli/main.py
@@ -2055,7 +2055,7 @@ class CliArgs:
         '-d',
         type=str,
         default='',
-        envvar='DSN',
+        envvar='MYSQL_DSN',
         help='DSN alias configured in the ~/.myclirc file, or a full DSN.',
     )
     list_dsn: bool = clickdc.option(
@@ -2343,6 +2343,16 @@ def click_entrypoint(
         )
         if not cli_args.socket:
             cli_args.socket = os.environ['MYSQL_UNIX_PORT']
+
+    if 'DSN' in os.environ:
+        # deprecated 2026-03
+        click.secho(
+            "The DSN environment variable is deprecated in favor of MYSQL_DSN.  Support for DSN will be removed in a future release.",
+            err=True,
+            fg="red",
+        )
+        if not cli_args.dsn:
+            cli_args.dsn = os.environ['DSN']
 
     # Choose which ever one has a valid value.
     database = cli_args.dbname or cli_args.database